requirements.txt
azure-identity==1.15.0
azure-mgmt-resource==23.0.1